// frontend/sanity.types.ts
export type SanityImageCrop = {
  _type: 'sanity.imageCrop'
  top?: number
  bottom?: number
  left?: number
  right?: number
}

export type SanityImageHotspot = {
  _type: 'sanity.imageHotspot'
  x?: number
  y?: number
  height?: number
  width?: number
}
// ...
export declare const internalGroqTypeReferenceTo: unique symbol
// Source: ../frontend/src/sanity/lib/queries.ts
// Variable: getAllDisciplinesQuery
// Query: *[_type == "disciplines"]{    ...,    "image": mainImage.asset->url  }
export type GetAllDisciplinesQueryResult = Array<{
  _id: string
  _type: 'disciplines'
  _createdAt: string
  _updatedAt: string
  _rev: string
  title?: string
  description?: string
  color?: string
  mainImage?: {
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    media?: unknown
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
  }
  challenges?: Array<{
    _ref: string
    _type: 'reference'
    _weak?: boolean
    _key: string
    [internalGroqTypeReferenceTo]?: 'challenge'
  }>
  image: string | null
}>